+++ module/modsupport/token-action-hud.js
@@ -1,18 +1,38 @@
-import { rollAttack, rollPower } from '../torgchecks.js';
+import { rollAttack, rollPower, rollAttribute, rollSkill } from '../torgchecks.js';
 
 // LEFT CLICK to perform ACTION
 // RIGHT CLICK to open Item Sheet (ignored on Skills)
 
 export default async function setupTokenActionHud(coreModule) {
 
-  /* ACTIONS */
-  const SKILLS_ID = 'skills';
-  const ATTACK_ID = 'combat';
-  const POWERS_ID = 'powers';
+  const ATTRIBUTES_ID = "attributesid";
+  const SKILLS_ID = 'skillsid';
+  const ATTACK_ID = 'combatid';
+  const POWERS_ID = 'powersid';
+  const GEAR_ID = 'gearid';   // since "gear" is the name of an Item group
+  const CONDITION_ID = 'conditionsid';
 
+  const ACTION_ATTRIBUTE = 'attribute';
   const ACTION_SKILL = 'skill';
   const ACTION_ATTACK = 'attack';
   const ACTION_POWER = 'power';
+  const ACTION_GEAR = 'gear';
+  const ACTION_CONDITION = 'conditions';
+
+  let GROUP = {
+    attributes: { id: ATTRIBUTES_ID, name: "torgeternity.sheetLabels.attributes", type: "system" },
+    skillsCombat: { id: 'skills_combat', name: "torgeternity.sheetLabels.combatSkills", type: "system" },
+    skillsInteraction: { id: 'skills_interaction', name: "torgeternity.sheetLabels.interactionSkills", type: "system" },
+    skillsOther: { id: 'skills_other', name: "torgeternity.sheetLabels.otherSkills", type: "system" },
+    combat: { id: ATTACK_ID, name: "torgeternity.sheetLabels.attacks", type: "system" },
+    powers: { id: POWERS_ID, name: "torgeternity.sheetLabels.powers", type: "system" },
+    gear: { id: GEAR_ID, name: "torgeternity.sheetLabels.gear", type: "system" },
+    conditions: { id: CONDITION_ID, name: "torgeternity.sheetLabels.conditions", type: "system" },
+  }
+  for (const key of Object.keys(CONFIG.Item.typeLabels)) {
+    if (key === 'base') continue;
+    GROUP[key] = { id: key, name: CONFIG.Item.typeLabels[key], type: "system" };
+  }
 
   class MyActionHandler extends coreModule.api.ActionHandler {
 
@@ -29,33 +49,52 @@ export default async function setupTokenActionHud(coreModule) {
       const actor = this.actor;
       if (!actor) return;
 
-      if (actor.type !== 'stormknight') return;
+      if (actor.type !== 'stormknight' && actor.type !== 'threat') return;
 
+      await this.#getAttributes(actor, tokenId, { id: ATTRIBUTES_ID, type: 'system' })
       await this.#getSkills(actor, tokenId, { id: SKILLS_ID, type: 'system' })
       await this.#getPowers(actor, tokenId, { id: POWERS_ID, type: 'system' })
       await this.#getAttacks(actor, tokenId, { id: ATTACK_ID, type: 'system' })
+      await this.#getGear(actor, tokenId, { id: GEAR_ID, type: 'system' })
+      await this.#getConditions(actor, tokenId, { id: CONDITION_ID, type: 'system' })
 
       //if (settings.get("showHudTitle")) result.hudTitle = token.name;
     }
 
+    async #getAttributes(actor, tokenId, parent) {
+      const actions = Object.entries(actor.system.attributes).map(([key, attribute]) => {
+        return {
+          id: key,
+          name: game.i18n.localize(`torgeternity.attributes.${key}`) + ` (${attribute.value})`,
+          encodedValue: [ACTION_ATTRIBUTE, actor.id, tokenId, key].join(this.delimiter),
+        }
+      })
+      const subcat = { id: `${parent.id}-${ATTRIBUTES_ID}`, name: coreModule.api.Utils.i18n('torgeternity.sheetLabels.attributes') };
+      this.addGroup(subcat, parent);
+      this.addActions(actions, subcat);
+    }
+
     async #getSkills(actor, tokenId, parent) {
+      const showUskilled = game.settings.get('torgeternity', 'tahShowUnskilled');
       const allSkills = Object.entries(actor.system.skills)
+        .filter(([key, skill]) => showUskilled || (skill.adds))
         .map(([key, skill]) => {
           return {
             id: key,
-            name: game.i18n.localize(`torgeternity.skills.${key}`),
-            groupName: skill.groupName,
+            name: game.i18n.localize(`torgeternity.skills.${key}`) + ` (${skill.value || '-'})`,
+            groupName: skill.groupName, // for local filtering
             encodedValue: [ACTION_SKILL, actor.id, tokenId, key].join(this.delimiter),
             //img: 'systems/torgeternity/images/icons/custom-skills.webp',
             //img: coreModule.api.Utils.getImage(skill),
             //tooltip: { content: skill.system.description },
-            system: skill
+            system: skill,
+            cssClass: (skill.value === 0) ? "unskilled" : "",
           }
         }).concat(actor.itemTypes.customSkill.map(skill => {
           return {
             id: skill.id,
-            name: skill.name,  // already in the local language
-            groupName: 'other',
+            name: skill.name + ` (${skill.value || '-'})`,  // already in the local language
+            groupName: 'other', // for local filtering
             encodedValue: [ACTION_SKILL, actor.id, tokenId, skill.id].join(this.delimiter),
             system: skill
           }
@@ -66,14 +105,14 @@ export default async function setupTokenActionHud(coreModule) {
       function skillGroup(main, groupName, label) {
         const actions = allSkills.filter(skill => skill.groupName === groupName);
         if (actions.length) {
-          const subcat = { id: `skill-${parent.id}-${groupName}`, name: coreModule.api.Utils.i18n(label), type: 'system-derived' };
+          const subcat = { id: `skills_${groupName}`, name: coreModule.api.Utils.i18n(label) };
           main.addGroup(subcat, parent);
           main.addActions(actions, subcat);
         }
       }
-      skillGroup(this, 'combat', 'torgeternity.sheetLabels.combatSkills', false);
-      skillGroup(this, 'interaction', 'torgeternity.sheetLabels.interactionSkills', true);
-      skillGroup(this, 'other', 'torgeternity.sheetLabels.otherSkills', false);
+      skillGroup(this, 'combat', 'torgeternity.sheetLabels.combatSkills');
+      skillGroup(this, 'interaction', 'torgeternity.sheetLabels.interactionSkills');
+      skillGroup(this, 'other', 'torgeternity.sheetLabels.otherSkills');
     }
 
     async #createList(parent, actor, tokenId, actionId, items, label, selectedfunc = undefined) {
@@ -96,22 +135,51 @@ export default async function setupTokenActionHud(coreModule) {
           });
 
       if (actions.length) {
-        const subcat = { id: `${parent.id}-${label}`, name: coreModule.api.Utils.i18n(label), type: 'system-derived' };
+        const subcat = { id: `${parent.id}-${label}`, name: coreModule.api.Utils.i18n(`torgeternity.sheetLabels.${label}`) };
         this.addGroup(subcat, parent);
         this.addActions(actions, subcat);
       }
     }
 
     async #getPowers(actor, tokenId, parent) {
-      await this.#createList(parent, actor, tokenId, ACTION_POWER, actor.itemTypes.spell, 'Spells');
-      await this.#createList(parent, actor, tokenId, ACTION_POWER, actor.itemTypes.miracle, 'Miracles');
-      await this.#createList(parent, actor, tokenId, ACTION_POWER, actor.itemTypes.psionicpower, 'Psionics');
+      await this.#createList(parent, actor, tokenId, ACTION_POWER, actor.itemTypes.spell, 'spells');
+      await this.#createList(parent, actor, tokenId, ACTION_POWER, actor.itemTypes.miracle, 'miracles');
+      await this.#createList(parent, actor, tokenId, ACTION_POWER, actor.itemTypes.psionicpower, 'psionicPowers');
     }
 
     async #getAttacks(actor, tokenId, parent) {
-      await this.#createList(parent, actor, tokenId, ACTION_ATTACK, actor.itemTypes.meleeweapon, 'Melee');
-      await this.#createList(parent, actor, tokenId, ACTION_ATTACK, actor.itemTypes.missileweapon, 'Missile');
-      await this.#createList(parent, actor, tokenId, ACTION_ATTACK, actor.itemTypes.firearm, 'Firearm');
+      await this.#createList(parent, actor, tokenId, ACTION_ATTACK, actor.itemTypes.meleeweapon, 'meleeWeapons');
+      await this.#createList(parent, actor, tokenId, ACTION_ATTACK, actor.itemTypes.missileweapon, 'missileWeapons');
+      await this.#createList(parent, actor, tokenId, ACTION_ATTACK, actor.itemTypes.firearm, 'firearms');
+      await this.#createList(parent, actor, tokenId, ACTION_ATTACK, actor.itemTypes.heavyweapon, 'heavyWeapons');
+      await this.#createList(parent, actor, tokenId, ACTION_ATTACK, actor.itemTypes["specialability-rollable"], 'specialabilities');
+      await this.#createList(parent, actor, tokenId, ACTION_ATTACK, actor.itemTypes.customAttack, 'customAttacks');
+    }
+
+    async #getGear(actor, tokenId, parent) {
+      await this.#createList(parent, actor, tokenId, ACTION_GEAR, [...actor.itemTypes.armor, ...actor.itemTypes.shield], 'armorAndShields');
+      await this.#createList(parent, actor, tokenId, ACTION_GEAR, actor.itemTypes.gear, 'generalGear');
+      await this.#createList(parent, actor, tokenId, ACTION_GEAR, actor.itemTypes.implant, 'implants');
+      await this.#createList(parent, actor, tokenId, ACTION_GEAR, actor.itemTypes.currency, 'currencies');
+      await this.#createList(parent, actor, tokenId, ACTION_GEAR, actor.itemTypes.eternityshard, 'eternityShard');
+    }
+
+    async #getConditions(actor, tokenId, parent) {
+      const actions = CONFIG.torgeternity.statusEffects.map(status => {
+        const result = {
+          id: status.id,
+          name: game.i18n.localize(status.name),
+          encodedValue: [ACTION_CONDITION, actor.id, tokenId, status.id].join(this.delimiter),
+          img: coreModule.api.Utils.getImage(status),
+          cssClass: `toggle${actor.statuses.has(status.id) ? " active" : ""}`,
+        }
+        const tooltip = actor.allApplicableEffects().find(effect => effect.statuses?.has(status.id))?.description;
+        if (tooltip) result.tooltip = { content: tooltip };
+        return result;
+      })
+      const subcat = { id: `${parent.id}-${CONDITION_ID}`, name: coreModule.api.Utils.i18n('torgeternity.sheetLabels.conditions') };
+      this.addGroup(subcat, parent);
+      this.addActions(actions, subcat);
     }
   } // MyActionHandler
 
@@ -132,33 +200,68 @@ export default async function setupTokenActionHud(coreModule) {
       const actor = coreModule.api.Utils.getActor(actorId, tokenId);
 
       switch (macroType) {
+
+        case ACTION_ATTRIBUTE:
+          // ActorSheet: skillRoll
+          if (this.isRenderItem()) return;
+          rollAttribute(this.actor, actionId);
+          break;
+
         case ACTION_SKILL:
           // ActorSheet: skillRoll
-          if (this.isRightClick) return;
-          game.torgeternity.rollSkillMacro(actionId, this.action.system.baseAttribute, /*isInteraction*/ this.action.system.groupName === 'interaction');
+          if (this.isRenderItem()) return;
+          rollSkill(this.actor, actionId);
           break;
+
         case ACTION_POWER:
           {
             // Sheet: itemPowerRoll
             const item = actor.items.get(actionId);
             if (item) {
-              if (this.isRightClick)
+              if (this.isRenderItem())
                 item.sheet.render({ force: true })
               else
                 rollPower(actor, item);
             }
           }
           break;
+
         case ACTION_ATTACK:
           // Sheet: onAttackRoll
           {
             const item = actor.items.get(actionId);
             if (item) {
-              if (this.isRightClick)
+              if (this.isRenderItem())
                 item.sheet.render({ force: true })
               else
                 rollAttack(actor, item);
             }
+          }
+          break;
+
+        case ACTION_GEAR:
+          {
+            const item = actor.items.get(actionId);
+            if (item) {
+              if (this.isRenderItem())
+                item.sheet.render({ force: true })
+              else
+                switch (item.type) {
+                  case 'eternityshard':
+                    // Actor Sheet: onTappingRoll
+                    break;
+                }
+            }
+          }
+          break;
+
+        case ACTION_CONDITION:
+          {
+            if (this.isRenderItem()) {
+              const effect = actor.allApplicableEffects().find(effect => effect.statuses.has(actionId));
+              if (effect) effect.sheet.render({ force: true }); // TODO
+            } else
+              this.actor.toggleStatusEffect(actionId);
           }
           break;
       }
@@ -173,8 +276,8 @@ export default async function setupTokenActionHud(coreModule) {
 
   class MySystemManager extends coreModule.api.SystemManager {
     /** @override */
-    getActionHandler(categoryManager) {
-      return new MyActionHandler(categoryManager)
+    getActionHandler() {
+      return new MyActionHandler()
     }
 
     /** @override */
@@ -189,66 +292,95 @@ export default async function setupTokenActionHud(coreModule) {
     }
 
     /** @override */
-    /*registerSettings (onChangeFunction) {
-        systemSettings.register(onChangeFunction)
-    }*/
+    registerSettings(onChangeFunction) {
+      game.settings.register('torgeternity', 'tahShowUnskilled', {
+        name: game.i18n.localize('torgeternity.settingMenu.tokenActionHud.showUnskilled.name'),
+        hint: game.i18n.localize('torgeternity.settingMenu.tokenActionHud.showUnskilled.hint'),
+        scope: 'client',
+        config: true,
+        type: Boolean,
+        default: false,
+        onChange: value => onChangeFunction(value)
+      })
+    }
 
     async registerDefaults() {
 
-      const SKILLS_NAME = game.i18n.localize('torgeternity.sheetLabels.skills');
-      const COMBAT_NAME = game.i18n.localize('torgeternity.sheetLabels.attacks');
-      const POWERS_NAME = game.i18n.localize('torgeternity.sheetLabels.powers');
+      const groups = GROUP;
+      Object.values(groups).forEach(group => {
+        group.name = game.i18n.localize(group.name);
+        group.listName = `Group: ${group.name}`;
+      });
 
       const DEFAULTS = {
         layout: [
           {
-            nestId: SKILLS_ID + "-title",
-            id: SKILLS_ID + "-title",
-            name: SKILLS_NAME,
+            nestId: ATTRIBUTES_ID,
+            id: ATTRIBUTES_ID,
+            name: game.i18n.localize('torgeternity.sheetLabels.attributes'),
             type: 'system',
             groups: [
-              {
-                nestId: 'skills-title_skills',
-                id: SKILLS_ID,
-                name: SKILLS_NAME,
-                type: 'system'
-              }
+              { ...groups.attributes, nestId: `attributes_attribute` },
             ]
           },
           {
-            nestId: ATTACK_ID + "-title",
-            id: ATTACK_ID + "-title",
-            name: COMBAT_NAME,
+            nestId: SKILLS_ID,
+            id: SKILLS_ID,
+            name: game.i18n.localize('torgeternity.sheetLabels.skills'),
             type: 'system',
             groups: [
-              {
-                nestId: 'combat-title_combat',
-                id: ATTACK_ID,
-                name: COMBAT_NAME,
-                type: 'system'
-              }
+              { ...groups.skillsCombat, nestId: "skills_combat" },
+              { ...groups.skillsInteraction, nestId: "skills_interaction" },
+              { ...groups.skillsOther, nestId: "skills_other" },
             ]
           },
           {
-            nestId: POWERS_ID + "-title",
-            id: POWERS_ID + "-title",
-            name: POWERS_NAME,
+            nestId: ATTACK_ID,
+            id: ATTACK_ID,
+            name: game.i18n.localize('torgeternity.sheetLabels.attacks'),
             type: 'system',
             groups: [
-              {
-                nestId: 'powers-title_powers',
-                id: POWERS_ID,
-                name: POWERS_NAME,
-                type: 'system'
-              }
+              { ...groups.meleeweapon, nestId: "attack_melee" },
+              { ...groups.missileweapon, nestId: "attack_missile" },
+              { ...groups.firearm, nestId: "attack_firearm" },
+            ]
+          },
+          {
+            nestId: POWERS_ID,
+            id: POWERS_ID,
+            name: game.i18n.localize('torgeternity.sheetLabels.powers'),
+            type: 'system',
+            groups: [
+              { ...groups.spell, nestId: "power_spell" },
+              { ...groups.miracle, nestId: "power_miracle" },
+              { ...groups.pisonicpower, nestId: "power_psionic" }
+            ]
+          },
+          {
+            nestId: GEAR_ID,
+            id: GEAR_ID,
+            name: game.i18n.localize('torgeternity.sheetLabels.gear'),
+            type: 'system',
+            groups: [
+              { ...groups.armor, nestId: "gear_armorAndShields" },
+              { ...groups.shield, nestId: "gear_shield" },
+              { ...groups.gear, nestId: "gear_generalGear" },
+              { ...groups.implant, nestId: "gear_implants" },
+              { ...groups.currency, nestId: "gear_currencies" },
+              { ...groups.eternityshard, nestId: "gear_eternityshard" },
+            ]
+          },
+          {
+            nestId: CONDITION_ID,
+            id: CONDITION_ID,
+            name: game.i18n.localize('torgeternity.sheetLabels.conditions'),
+            type: 'system',
+            groups: [
+              { ...groups.conditions, nestId: "condition_condition" },
             ]
           },
         ],
-        groups: [
-          { id: ATTACK_ID, name: COMBAT_NAME, type: 'system' },
-          { id: SKILLS_ID, name: SKILLS_NAME, type: 'system' },
-          { id: POWERS_ID, name: POWERS_NAME, type: 'system' },
-        ]
+        groups: Object.values(groups)
       }
 
       // HUD CORE v1.2 wants us to return the DEFAULTS