+++ Changelog.md
@@ -24,6 +24,7 @@
   - Item Piles Currencies are determined by looking for all Currency Items in the world sidebar and in compendiums.
   - **TODO**: Please raise an issue if there are any improvements to the default configuration which you find useful.
 - Add some basic support for **Token Action HUD** providing Skills, Attacks and Powers in the HUD.
+  - A game setting determines whether unskilled skills appear in the HUD's Skills list.
 
 ## 13.18.3 - Resetting Possibilities & Tapping Eternity Shards